move clips page overlay to display over multiplayer
